@@ -18,13 +18,14 @@ package com.bwaim.musicalstructure.ui;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.GridView;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
 
 import com.bwaim.musicalstructure.R;
 import com.bwaim.musicalstructure.adapters.ArtistAdapter;
@@ -94,7 +95,7 @@ public class ArtistFragment extends Fragment {
     }
 
     @Override
-    public void onAttach(Context context) {
+    public void onAttach(@NonNull Context context) {
         super.onAttach(context);
         if (context instanceof OnArtistSelectedListener) {
             mListener = (OnArtistSelectedListener) context;
